lambdas/wx_poller/requirements.txt
requests==2.32.3
boto3
Pillow==10.4.0
